+++ Source/Core/Image/XYZColor.cpp
@@ -39,10 +39,10 @@ inline kvs::Real32 ToLinear( const kvs::Real32 C )
 /*===========================================================================*/
 inline kvs::Vec3 sRGB2RGB( const kvs::Vec3 sRGB )
 {
-    const kvs::Real32 R = ToLinear( sRGB[0] );
-    const kvs::Real32 G = ToLinear( sRGB[1] );
-    const kvs::Real32 B = ToLinear( sRGB[2] );
-    return kvs::Vec3( R, G, B );
+    const auto R = ToLinear( sRGB[0] );
+    const auto G = ToLinear( sRGB[1] );
+    const auto B = ToLinear( sRGB[2] );
+    return { R, G, B };
 }
 
 /*===========================================================================*/
@@ -54,61 +54,40 @@ inline kvs::Vec3 sRGB2RGB( const kvs::Vec3 sRGB )
 /*===========================================================================*/
 inline kvs::Vec3 RGB2sRGB( const kvs::Vec3 RGB )
 {
-    const kvs::Real32 sR = ToStandard( RGB[0] );
-    const kvs::Real32 sG = ToStandard( RGB[1] );
-    const kvs::Real32 sB = ToStandard( RGB[2] );
-    return kvs::Vec3( sR, sG, sB );
+    const auto sR = ToStandard( RGB[0] );
+    const auto sG = ToStandard( RGB[1] );
+    const auto sB = ToStandard( RGB[2] );
+    return { sR, sG, sB };
 }
 
 inline kvs::XYZColor RGB2XYZ( const kvs::RGBColor& rgb )
 {
-    const kvs::Vec3 sRGB = rgb.toVec3();
-    const kvs::Vec3 RGB = sRGB2RGB( sRGB );
-    const kvs::Mat3 M(
+    const auto M = kvs::Mat3{
         0.412391f, 0.357584f, 0.180481f,
         0.212639f, 0.715169f, 0.072192f,
-        0.019331f, 0.119195f, 0.950532f );
-    const kvs::Vec3 XYZ = M * RGB;
-    return kvs::XYZColor( XYZ[0], XYZ[1], XYZ[2] );
+        0.019331f, 0.119195f, 0.950532f };
+    const auto sRGB = rgb.toVec3();
+    const auto RGB = sRGB2RGB( sRGB );
+    const auto XYZ = M * RGB;
+    return { XYZ[0], XYZ[1], XYZ[2] };
 }
 
 inline kvs::RGBColor XYZ2RGB( const kvs::XYZColor& xyz )
 {
-    const kvs::Mat3 Minv(
+    const auto Minv = kvs::Mat3{
          3.240970f, -1.537383f, -0.498611f,
         -0.969244f,  1.875968f,  0.041555f,
-         0.055630f, -0.203977f,  1.056972f );
-    const kvs::Vec3 RGB = Minv * xyz.toVec3();
-    const kvs::Vec3 sRGB = RGB2sRGB( RGB );
-    return kvs::RGBColor( sRGB );
+         0.055630f, -0.203977f,  1.056972f };
+    const auto RGB = Minv * xyz.toVec3();
+    const auto sRGB = RGB2sRGB( RGB );
+    return { sRGB };
 }
 
-}
+} // end of namespace
 
 
 namespace kvs
 {
-
-XYZColor::XYZColor( kvs::Real32 x, kvs::Real32 y, kvs::Real32 z ):
-    m_x( x ),
-    m_y( y ),
-    m_z( z )
-{
-}
-
-XYZColor::XYZColor( const kvs::Vec3& xyz ):
-    m_x( xyz[0] ),
-    m_y( xyz[1] ),
-    m_z( xyz[2] )
-{
-}
-
-XYZColor::XYZColor( const kvs::XYZColor& xyz ):
-    m_x( xyz.x() ),
-    m_y( xyz.y() ),
-    m_z( xyz.z() )
-{
-}
 
 XYZColor::XYZColor( const kvs::RGBColor& rgb )
 {