+++ Source/Core/Image/XYZColor.h
@@ -23,14 +23,15 @@ class LabColor;
 class XYZColor
 {
 private:
-    kvs::Real32 m_x; ///< [0-1]
-    kvs::Real32 m_y; ///< [0-1]
-    kvs::Real32 m_z; ///< [0-1]
+    kvs::Real32 m_x = 0.0f; ///< [0-1]
+    kvs::Real32 m_y = 0.0f; ///< [0-1]
+    kvs::Real32 m_z = 0.0f; ///< [0-1]
 
 public:
-    XYZColor( kvs::Real32 x, kvs::Real32 y, kvs::Real32 z );
-    XYZColor( const kvs::Vec3& xyz );
-    XYZColor( const kvs::XYZColor& xyz );
+    XYZColor() = default;
+    XYZColor( kvs::Real32 x, kvs::Real32 y, kvs::Real32 z ): m_x( x ), m_y( y ), m_z( z ) {}
+    XYZColor( const kvs::Vec3& xyz ): m_x( xyz[0] ), m_y( xyz[1] ), m_z( xyz[2] ) {}
+    XYZColor( const kvs::XYZColor& xyz ): m_x( xyz.x() ), m_y( xyz.y() ), m_z( xyz.z() ) {}
     XYZColor( const kvs::RGBColor& rgb );
 
     void set( kvs::Real32 x, kvs::Real32 y, kvs::Real32 z ) { m_x = x; m_y = y; m_z = z; }