+++ Source/Core/Visualization/Mapper/ColorMap.cpp
@@ -7,6 +7,7 @@
 #include "ColorMap.h"
 #include <kvs/Assert>
 #include <kvs/RGBColor>
+#include <kvs/HCLColor>
 #include <kvs/HSLColor>
 #include <kvs/HSVColor>
 #include <kvs/LabColor>
@@ -20,41 +21,6 @@ namespace
 // Default values.
 const size_t NumberOfChannels = 3;
 kvs::ColorMap::ColorMapFunction DefaultColorMap = kvs::ColorMap::BrewerRdBu;
-
-/*===========================================================================*/
- /**
-  *  @brief  Returns a piecewise-linearly interpolated color map.
-  *  @param  colors [in] color list
-  *  @param  resolution [in] resolution of the color map
-  *  @param  space [in] interpolation color space
-  *  @return color map interpolated in the specified color space
-  */
- /*===========================================================================*/
-inline kvs::ColorMap Interpolate(
-    const std::list<kvs::RGBColor>& colors,
-    const size_t resolution,
-    kvs::ColorMap::ColorSpace space )
-{
-    const size_t ncolors = colors.size();
-    const float stride = 1.0f / ( ncolors - 1 );
-
-    kvs::ColorMap cmap( resolution );
-    cmap.addPoint( 0.0f, colors.front() ); // start point
-    auto color = colors.begin(); color++;
-    auto end = colors.end(); end--;
-    size_t index = 1;
-    while ( color != end )
-    {
-        const float value = kvs::Math::Round( resolution * stride * index );
-        cmap.addPoint( value, *color );
-        color++; index++;
-    }
-    cmap.addPoint( float( resolution - 1 ), colors.back() ); // end point
-    cmap.setColorSpace( space );
-    cmap.create();
-
-    return cmap;
-}
 
 }
 
@@ -76,10 +42,18 @@ void ColorMap::SetDefaultColorMap( ColorMapFunction func )
 /*===========================================================================*/
 kvs::ColorMap ColorMap::Rainbow( const size_t resolution )
 {
-    std::list<kvs::RGBColor> colors;
-    colors.push_back( kvs::HSVColor(   0.0f, 0.0f, 0.0f ) );
-    colors.push_back( kvs::RGBColor( 240.0f, 0.0f, 0.0f ) );
-    return ::Interpolate( colors, resolution, kvs::ColorMap::HSVSpace );
+    const auto hue_min = 0.0f;
+    const auto hue_max = 240.0f / 360.0f;
+    std::list<kvs::RGBColor> colors = {
+        kvs::HSVColor( hue_max, 1.0f, 1.0f ), // blue
+        kvs::HSVColor( hue_min, 1.0f, 1.0f )  // red
+    };
+
+    kvs::ColorMap cmap( resolution );
+    cmap.setPoints( colors );
+    cmap.setColorSpaceToHSV();
+    cmap.create();
+    return cmap;
 }
 
 /*===========================================================================*/
@@ -98,10 +72,16 @@ kvs::ColorMap ColorMap::Rainbow( const size_t resolution )
 /*===========================================================================*/
 kvs::ColorMap ColorMap::CoolWarm( const size_t resolution )
 {
-    std::list<kvs::RGBColor> colors;
-    colors.push_back( kvs::RGBColor(  59, 76, 192 ) );
-    colors.push_back( kvs::RGBColor( 180,  4,  38 ) );
-    return ::Interpolate( colors, resolution, kvs::ColorMap::MshSpace );
+    std::list<kvs::RGBColor> colors = {
+        {  59, 76, 192 },
+        { 180,  4,  38 }
+    };
+
+    kvs::ColorMap cmap( resolution );
+    cmap.setPoints( colors );
+    cmap.setColorSpaceToMsh();
+    cmap.create();
+    return cmap;
 }
 
 /*===========================================================================*/
@@ -119,20 +99,26 @@ kvs::ColorMap ColorMap::CoolWarm( const size_t resolution )
 /*===========================================================================*/
 kvs::ColorMap ColorMap::BrewerBrBG( const size_t resolution )
 {
-    std::list<kvs::RGBColor> colors;
-    colors.push_back( kvs::RGBColor(  84,  48,   5 ) );
-    colors.push_back( kvs::RGBColor( 140,  81,  10 ) );
-    colors.push_back( kvs::RGBColor( 191, 129,  45 ) );
-    colors.push_back( kvs::RGBColor( 223, 194, 125 ) );
-    colors.push_back( kvs::RGBColor( 246, 232, 195 ) );
-    colors.push_back( kvs::RGBColor( 245, 245, 245 ) );
-    colors.push_back( kvs::RGBColor( 199, 234, 229 ) );
-    colors.push_back( kvs::RGBColor( 128, 205, 193 ) );
-    colors.push_back( kvs::RGBColor(  53, 151, 143 ) );
-    colors.push_back( kvs::RGBColor(   1, 102,  94 ) );
-    colors.push_back( kvs::RGBColor(   0,  60,  48 ) );
+    std::list<kvs::RGBColor> colors = {
+        {  84,  48,   5 },
+        { 140,  81,  10 },
+        { 191, 129,  45 },
+        { 223, 194, 125 },
+        { 246, 232, 195 },
+        { 245, 245, 245 },
+        { 199, 234, 229 },
+        { 128, 205, 193 },
+        {  53, 151, 143 },
+        {   1, 102,  94 },
+        {   0,  60,  48 }
+    };
     colors.reverse();
-    return ::Interpolate( colors, resolution, kvs::ColorMap::LabSpace );
+
+    kvs::ColorMap cmap( resolution );
+    cmap.setPoints( colors );
+    cmap.setColorSpaceToLab();
+    cmap.create();
+    return cmap;
 }
 
 /*===========================================================================*/
@@ -150,20 +136,26 @@ kvs::ColorMap ColorMap::BrewerBrBG( const size_t resolution )
 /*===========================================================================*/
 kvs::ColorMap ColorMap::BrewerPiYG( const size_t resolution )
 {
-    std::list<kvs::RGBColor> colors;
-    colors.push_back( kvs::RGBColor( 142,   1,  82 ) );
-    colors.push_back( kvs::RGBColor( 197,  27, 125 ) );
-    colors.push_back( kvs::RGBColor( 222, 119, 174 ) );
-    colors.push_back( kvs::RGBColor( 241, 182, 218 ) );
-    colors.push_back( kvs::RGBColor( 253, 224, 239 ) );
-    colors.push_back( kvs::RGBColor( 247, 247, 247 ) );
-    colors.push_back( kvs::RGBColor( 230, 245, 208 ) );
-    colors.push_back( kvs::RGBColor( 184, 225, 134 ) );
-    colors.push_back( kvs::RGBColor( 127, 188, 134 ) );
-    colors.push_back( kvs::RGBColor(  77, 146,  33 ) );
-    colors.push_back( kvs::RGBColor(  39, 100,  25 ) );
+    std::list<kvs::RGBColor> colors = {
+        { 142,   1,  82 },
+        { 197,  27, 125 },
+        { 222, 119, 174 },
+        { 241, 182, 218 },
+        { 253, 224, 239 },
+        { 247, 247, 247 },
+        { 230, 245, 208 },
+        { 184, 225, 134 },
+        { 127, 188, 134 },
+        {  77, 146,  33 },
+        {  39, 100,  25 }
+    };
     colors.reverse();
-    return ::Interpolate( colors, resolution, kvs::ColorMap::LabSpace );
+
+    kvs::ColorMap cmap( resolution );
+    cmap.setPoints( colors );
+    cmap.setColorSpaceToLab();
+    cmap.create();
+    return cmap;
 }
 
 /*===========================================================================*/
@@ -181,20 +173,26 @@ kvs::ColorMap ColorMap::BrewerPiYG( const size_t resolution )
 /*===========================================================================*/
 kvs::ColorMap ColorMap::BrewerPRGn( const size_t resolution )
 {
-    std::list<kvs::RGBColor> colors;
-    colors.push_back( kvs::RGBColor(  64,   0,  75 ) );
-    colors.push_back( kvs::RGBColor( 118,  42, 131 ) );
-    colors.push_back( kvs::RGBColor( 153, 112, 171 ) );
-    colors.push_back( kvs::RGBColor( 194, 165, 207 ) );
-    colors.push_back( kvs::RGBColor( 231, 212, 232 ) );
-    colors.push_back( kvs::RGBColor( 247, 247, 247 ) );
-    colors.push_back( kvs::RGBColor( 217, 240, 211 ) );
-    colors.push_back( kvs::RGBColor( 166, 219, 160 ) );
-    colors.push_back( kvs::RGBColor(  90, 174,  97 ) );
-    colors.push_back( kvs::RGBColor(  27, 120,  55 ) );
-    colors.push_back( kvs::RGBColor(   0,  68,  27 ) );
+    std::list<kvs::RGBColor> colors = {
+        {  64,   0,  75 },
+        { 118,  42, 131 },
+        { 153, 112, 171 },
+        { 194, 165, 207 },
+        { 231, 212, 232 },
+        { 247, 247, 247 },
+        { 217, 240, 211 },
+        { 166, 219, 160 },
+        {  90, 174,  97 },
+        {  27, 120,  55 },
+        {   0,  68,  27 }
+    };
     colors.reverse();
-    return ::Interpolate( colors, resolution, kvs::ColorMap::LabSpace );
+
+    kvs::ColorMap cmap( resolution );
+    cmap.setPoints( colors );
+    cmap.setColorSpaceToLab();
+    cmap.create();
+    return cmap;
 }
 
 /*===========================================================================*/
@@ -212,20 +210,26 @@ kvs::ColorMap ColorMap::BrewerPRGn( const size_t resolution )
 /*===========================================================================*/
 kvs::ColorMap ColorMap::BrewerPuOr( const size_t resolution )
 {
-    std::list<kvs::RGBColor> colors;
-    colors.push_back( kvs::RGBColor( 127,  59,   8 ) );
-    colors.push_back( kvs::RGBColor( 179,  88,   6 ) );
-    colors.push_back( kvs::RGBColor( 224, 130,  20 ) );
-    colors.push_back( kvs::RGBColor( 253, 184,  99 ) );
-    colors.push_back( kvs::RGBColor( 254, 224, 182 ) );
-    colors.push_back( kvs::RGBColor( 247, 247, 247 ) );
-    colors.push_back( kvs::RGBColor( 216, 218, 235 ) );
-    colors.push_back( kvs::RGBColor( 178, 171, 210 ) );
-    colors.push_back( kvs::RGBColor( 128, 115, 172 ) );
-    colors.push_back( kvs::RGBColor(  84,  39, 136 ) );
-    colors.push_back( kvs::RGBColor(  45,   0,  75 ) );
+    std::list<kvs::RGBColor> colors = {
+        { 127,  59,   8 },
+        { 179,  88,   6 },
+        { 224, 130,  20 },
+        { 253, 184,  99 },
+        { 254, 224, 182 },
+        { 247, 247, 247 },
+        { 216, 218, 235 },
+        { 178, 171, 210 },
+        { 128, 115, 172 },
+        {  84,  39, 136 },
+        {  45,   0,  75 }
+    };
     colors.reverse();
-    return ::Interpolate( colors, resolution, kvs::ColorMap::LabSpace );
+
+    kvs::ColorMap cmap( resolution );
+    cmap.setPoints( colors );
+    cmap.setColorSpaceToLab();
+    cmap.create();
+    return cmap;
 }
 
 /*===========================================================================*/
@@ -243,20 +247,26 @@ kvs::ColorMap ColorMap::BrewerPuOr( const size_t resolution )
 /*===========================================================================*/
 kvs::ColorMap ColorMap::BrewerRdBu( const size_t resolution )
 {
-    std::list<kvs::RGBColor> colors;
-    colors.push_back( kvs::RGBColor( 103,   0,  31 ) );
-    colors.push_back( kvs::RGBColor( 178,  24,  43 ) );
-    colors.push_back( kvs::RGBColor( 214,  96,  77 ) );
-    colors.push_back( kvs::RGBColor( 244, 165, 130 ) );
-    colors.push_back( kvs::RGBColor( 253, 219, 199 ) );
-    colors.push_back( kvs::RGBColor( 247, 247, 247 ) );
-    colors.push_back( kvs::RGBColor( 209, 229, 240 ) );
-    colors.push_back( kvs::RGBColor( 146, 197, 222 ) );
-    colors.push_back( kvs::RGBColor(  67, 147, 195 ) );
-    colors.push_back( kvs::RGBColor(  33, 102, 172 ) );
-    colors.push_back( kvs::RGBColor(   5,  48,  97 ) );
+    std::list<kvs::RGBColor> colors = {
+        { 103,   0,  31 },
+        { 178,  24,  43 },
+        { 214,  96,  77 },
+        { 244, 165, 130 },
+        { 253, 219, 199 },
+        { 247, 247, 247 },
+        { 209, 229, 240 },
+        { 146, 197, 222 },
+        {  67, 147, 195 },
+        {  33, 102, 172 },
+        {   5,  48,  97 }
+    };
     colors.reverse();
-    return ::Interpolate( colors, resolution, kvs::ColorMap::LabSpace );
+
+    kvs::ColorMap cmap( resolution );
+    cmap.setPoints( colors );
+    cmap.setColorSpaceToLab();
+    cmap.create();
+    return cmap;
 }
 
 /*===========================================================================*/
@@ -274,20 +284,26 @@ kvs::ColorMap ColorMap::BrewerRdBu( const size_t resolution )
 /*===========================================================================*/
 kvs::ColorMap ColorMap::BrewerRdGy( const size_t resolution )
 {
-    std::list<kvs::RGBColor> colors;
-    colors.push_back( kvs::RGBColor( 103,   0,  31 ) );
-    colors.push_back( kvs::RGBColor( 178,  24,  43 ) );
-    colors.push_back( kvs::RGBColor( 214,  96,  77 ) );
-    colors.push_back( kvs::RGBColor( 244, 165, 130 ) );
-    colors.push_back( kvs::RGBColor( 253, 219, 199 ) );
-    colors.push_back( kvs::RGBColor( 255, 255, 255 ) );
-    colors.push_back( kvs::RGBColor( 224, 224, 224 ) );
-    colors.push_back( kvs::RGBColor( 186, 186, 186 ) );
-    colors.push_back( kvs::RGBColor( 135, 135, 135 ) );
-    colors.push_back( kvs::RGBColor(  77,  77,  77 ) );
-    colors.push_back( kvs::RGBColor(  26,  26,  26 ) );
+    std::list<kvs::RGBColor> colors = {
+        { 103,   0,  31 },
+        { 178,  24,  43 },
+        { 214,  96,  77 },
+        { 244, 165, 130 },
+        { 253, 219, 199 },
+        { 255, 255, 255 },
+        { 224, 224, 224 },
+        { 186, 186, 186 },
+        { 135, 135, 135 },
+        {  77,  77,  77 },
+        {  26,  26,  26 }
+    };
     colors.reverse();
-    return ::Interpolate( colors, resolution, kvs::ColorMap::LabSpace );
+
+    kvs::ColorMap cmap( resolution );
+    cmap.setPoints( colors );
+    cmap.setColorSpaceToLab();
+    cmap.create();
+    return cmap;
 }
 
 /*===========================================================================*/
@@ -305,20 +321,26 @@ kvs::ColorMap ColorMap::BrewerRdGy( const size_t resolution )
 /*===========================================================================*/
 kvs::ColorMap ColorMap::BrewerRdYlBu( const size_t resolution )
 {
-    std::list<kvs::RGBColor> colors;
-    colors.push_back( kvs::RGBColor( 165,   0,  38 ) );
-    colors.push_back( kvs::RGBColor( 215,  48,  39 ) );
-    colors.push_back( kvs::RGBColor( 244, 109,  67 ) );
-    colors.push_back( kvs::RGBColor( 253, 174,  97 ) );
-    colors.push_back( kvs::RGBColor( 254, 224, 144 ) );
-    colors.push_back( kvs::RGBColor( 255, 255, 191 ) );
-    colors.push_back( kvs::RGBColor( 224, 243, 248 ) );
-    colors.push_back( kvs::RGBColor( 171, 217, 233 ) );
-    colors.push_back( kvs::RGBColor( 116, 173, 209 ) );
-    colors.push_back( kvs::RGBColor(  69, 117, 180 ) );
-    colors.push_back( kvs::RGBColor(  49,  54, 149 ) );
+    std::list<kvs::RGBColor> colors = {
+        { 165,   0,  38 },
+        { 215,  48,  39 },
+        { 244, 109,  67 },
+        { 253, 174,  97 },
+        { 254, 224, 144 },
+        { 255, 255, 191 },
+        { 224, 243, 248 },
+        { 171, 217, 233 },
+        { 116, 173, 209 },
+        {  69, 117, 180 },
+        {  49,  54, 149 }
+    };
     colors.reverse();
-    return ::Interpolate( colors, resolution, kvs::ColorMap::LabSpace );
+
+    kvs::ColorMap cmap( resolution );
+    cmap.setPoints( colors );
+    cmap.setColorSpaceToLab();
+    cmap.create();
+    return cmap;
 }
 
 /*===========================================================================*/
@@ -336,20 +358,26 @@ kvs::ColorMap ColorMap::BrewerRdYlBu( const size_t resolution )
 /*===========================================================================*/
 kvs::ColorMap ColorMap::BrewerRdYlGn( const size_t resolution )
 {
-    std::list<kvs::RGBColor> colors;
-    colors.push_back( kvs::RGBColor( 165,   0,  38 ) );
-    colors.push_back( kvs::RGBColor( 215,  48,  39 ) );
-    colors.push_back( kvs::RGBColor( 244, 109,  67 ) );
-    colors.push_back( kvs::RGBColor( 253, 174,  97 ) );
-    colors.push_back( kvs::RGBColor( 254, 224, 139 ) );
-    colors.push_back( kvs::RGBColor( 255, 255, 191 ) );
-    colors.push_back( kvs::RGBColor( 217, 239, 139 ) );
-    colors.push_back( kvs::RGBColor( 166, 217, 106 ) );
-    colors.push_back( kvs::RGBColor( 102, 189,  99 ) );
-    colors.push_back( kvs::RGBColor(  26, 152,  80 ) );
-    colors.push_back( kvs::RGBColor(   0, 104,  55 ) );
+    std::list<kvs::RGBColor> colors = {
+        { 165,   0,  38 },
+        { 215,  48,  39 },
+        { 244, 109,  67 },
+        { 253, 174,  97 },
+        { 254, 224, 139 },
+        { 255, 255, 191 },
+        { 217, 239, 139 },
+        { 166, 217, 106 },
+        { 102, 189,  99 },
+        {  26, 152,  80 },
+        {   0, 104,  55 }
+    };
     colors.reverse();
-    return ::Interpolate( colors, resolution, kvs::ColorMap::LabSpace );
+
+    kvs::ColorMap cmap( resolution );
+    cmap.setPoints( colors );
+    cmap.setColorSpaceToLab();
+    cmap.create();
+    return cmap;
 }
 
 /*===========================================================================*/
@@ -367,20 +395,26 @@ kvs::ColorMap ColorMap::BrewerRdYlGn( const size_t resolution )
 /*===========================================================================*/
 kvs::ColorMap ColorMap::BrewerSpectral( const size_t resolution )
 {
-    std::list<kvs::RGBColor> colors;
-    colors.push_back( kvs::RGBColor( 158,   1,  66 ) );
-    colors.push_back( kvs::RGBColor( 213,  62,  79 ) );
-    colors.push_back( kvs::RGBColor( 244, 109,  67 ) );
-    colors.push_back( kvs::RGBColor( 253, 174,  97 ) );
-    colors.push_back( kvs::RGBColor( 254, 224, 139 ) );
-    colors.push_back( kvs::RGBColor( 255, 255, 191 ) );
-    colors.push_back( kvs::RGBColor( 230, 245, 152 ) );
-    colors.push_back( kvs::RGBColor( 171, 221, 164 ) );
-    colors.push_back( kvs::RGBColor( 102, 194, 165 ) );
-    colors.push_back( kvs::RGBColor(  50, 136, 189 ) );
-    colors.push_back( kvs::RGBColor(  94,  79, 162 ) );
+    std::list<kvs::RGBColor> colors = {
+        { 158,   1,  66 },
+        { 213,  62,  79 },
+        { 244, 109,  67 },
+        { 253, 174,  97 },
+        { 254, 224, 139 },
+        { 255, 255, 191 },
+        { 230, 245, 152 },
+        { 171, 221, 164 },
+        { 102, 194, 165 },
+        {  50, 136, 189 },
+        {  94,  79, 162 }
+    };
     colors.reverse();
-    return ::Interpolate( colors, resolution, kvs::ColorMap::LabSpace );
+
+    kvs::ColorMap cmap( resolution );
+    cmap.setPoints( colors );
+    cmap.setColorSpaceToLab();
+    cmap.create();
+    return cmap;
 }
 
 /*===========================================================================*/
@@ -392,19 +426,25 @@ kvs::ColorMap ColorMap::BrewerSpectral( const size_t resolution )
 /*===========================================================================*/
 kvs::ColorMap ColorMap::Viridis( const size_t resolution )
 {
-    std::list<kvs::RGBColor> colors;
-    colors.push_back( kvs::RGBColor( 253, 231,  37 ) );
-    colors.push_back( kvs::RGBColor( 181, 221,  43 ) );
-    colors.push_back( kvs::RGBColor( 109, 206,  88 ) );
-    colors.push_back( kvs::RGBColor(  53, 183, 120 ) );
-    colors.push_back( kvs::RGBColor(  30, 156, 137 ) );
-    colors.push_back( kvs::RGBColor(  37, 130, 142 ) );
-    colors.push_back( kvs::RGBColor(  49, 102, 141 ) );
-    colors.push_back( kvs::RGBColor(  62,  72, 136 ) );
-    colors.push_back( kvs::RGBColor(  71,  38, 118 ) );
-    colors.push_back( kvs::RGBColor(  68,   1,  84 ) );
+    std::list<kvs::RGBColor> colors = {
+        { 253, 231,  37 },
+        { 181, 221,  43 },
+        { 109, 206,  88 },
+        {  53, 183, 120 },
+        {  30, 156, 137 },
+        {  37, 130, 142 },
+        {  49, 102, 141 },
+        {  62,  72, 136 },
+        {  71,  38, 118 },
+        {  68,   1,  84 }
+    };
     colors.reverse();
-    return ::Interpolate( colors, resolution, kvs::ColorMap::LabSpace );
+
+    kvs::ColorMap cmap( resolution );
+    cmap.setPoints( colors );
+    cmap.setColorSpaceToLab();
+    cmap.create();
+    return cmap;
 }
 
 /*===========================================================================*/
@@ -416,19 +456,25 @@ kvs::ColorMap ColorMap::Viridis( const size_t resolution )
 /*===========================================================================*/
 kvs::ColorMap ColorMap::Plasma( const size_t resolution )
 {
-    std::list<kvs::RGBColor> colors;
-    colors.push_back( kvs::RGBColor( 239, 248,  33 ) );
-    colors.push_back( kvs::RGBColor( 252, 201,  38 ) );
-    colors.push_back( kvs::RGBColor( 250, 159,  58 ) );
-    colors.push_back( kvs::RGBColor( 236, 120,  83 ) );
-    colors.push_back( kvs::RGBColor( 215,  86, 108 ) );
-    colors.push_back( kvs::RGBColor( 188,  54, 133 ) );
-    colors.push_back( kvs::RGBColor( 154,  21, 158 ) );
-    colors.push_back( kvs::RGBColor( 112,   0, 168 ) );
-    colors.push_back( kvs::RGBColor(  68,   3, 158 ) );
-    colors.push_back( kvs::RGBColor(  12,   7, 134 ) );
+    std::list<kvs::RGBColor> colors = {
+        { 239, 248,  33 },
+        { 252, 201,  38 },
+        { 250, 159,  58 },
+        { 236, 120,  83 },
+        { 215,  86, 108 },
+        { 188,  54, 133 },
+        { 154,  21, 158 },
+        { 112,   0, 168 },
+        {  68,   3, 158 },
+        {  12,   7, 134 }
+    };
     colors.reverse();
-    return ::Interpolate( colors, resolution, kvs::ColorMap::LabSpace );
+
+    kvs::ColorMap cmap( resolution );
+    cmap.setPoints( colors );
+    cmap.setColorSpaceToLab();
+    cmap.create();
+    return cmap;
 }
 
 /*===========================================================================*/
@@ -440,19 +486,25 @@ kvs::ColorMap ColorMap::Plasma( const size_t resolution )
 /*===========================================================================*/
 kvs::ColorMap ColorMap::Inferno( const size_t resolution )
 {
-    std::list<kvs::RGBColor> colors;
-    colors.push_back( kvs::RGBColor( 252, 254, 164 ) );
-    colors.push_back( kvs::RGBColor( 247, 209,  60 ) );
-    colors.push_back( kvs::RGBColor( 251, 155,   6 ) );
-    colors.push_back( kvs::RGBColor( 237, 104,  37 ) );
-    colors.push_back( kvs::RGBColor( 205,  66,  71 ) );
-    colors.push_back( kvs::RGBColor( 164,  44,  96 ) );
-    colors.push_back( kvs::RGBColor( 118,  27, 109 ) );
-    colors.push_back( kvs::RGBColor(  72,  11, 106 ) );
-    colors.push_back( kvs::RGBColor(  25,  11,  62 ) );
-    colors.push_back( kvs::RGBColor(   0,   0,   3 ) );
+    std::list<kvs::RGBColor> colors = {
+        { 252, 254, 164 },
+        { 247, 209,  60 },
+        { 251, 155,   6 },
+        { 237, 104,  37 },
+        { 205,  66,  71 },
+        { 164,  44,  96 },
+        { 118,  27, 109 },
+        {  72,  11, 106 },
+        {  25,  11,  62 },
+        {   0,   0,   3 }
+    };
     colors.reverse();
-    return ::Interpolate( colors, resolution, kvs::ColorMap::LabSpace );
+
+    kvs::ColorMap cmap( resolution );
+    cmap.setPoints( colors );
+    cmap.setColorSpaceToLab();
+    cmap.create();
+    return cmap;
 }
 
 /*===========================================================================*/
@@ -464,19 +516,25 @@ kvs::ColorMap ColorMap::Inferno( const size_t resolution )
 /*===========================================================================*/
 kvs::ColorMap ColorMap::Magma( const size_t resolution )
 {
-    std::list<kvs::RGBColor> colors;
-    colors.push_back( kvs::RGBColor( 251, 252, 191 ) );
-    colors.push_back( kvs::RGBColor( 254, 201, 141 ) );
-    colors.push_back( kvs::RGBColor( 253, 149, 103 ) );
-    colors.push_back( kvs::RGBColor( 240,  96,  93 ) );
-    colors.push_back( kvs::RGBColor( 203,  62, 113 ) );
-    colors.push_back( kvs::RGBColor( 158,  46, 126 ) );
-    colors.push_back( kvs::RGBColor( 111,  30, 129 ) );
-    colors.push_back( kvs::RGBColor(  66,  15, 116 ) );
-    colors.push_back( kvs::RGBColor(  22,  14,  58 ) );
-    colors.push_back( kvs::RGBColor(   0,   0,   3 ) );
+    std::list<kvs::RGBColor> colors = {
+        { 251, 252, 191 },
+        { 254, 201, 141 },
+        { 253, 149, 103 },
+        { 240,  96,  93 },
+        { 203,  62, 113 },
+        { 158,  46, 126 },
+        { 111,  30, 129 },
+        {  66,  15, 116 },
+        {  22,  14,  58 },
+        {   0,   0,   3 }
+    };
     colors.reverse();
-    return ::Interpolate( colors, resolution, kvs::ColorMap::LabSpace );
+
+    kvs::ColorMap cmap( resolution );
+    cmap.setPoints( colors );
+    cmap.setColorSpaceToLab();
+    cmap.create();
+    return cmap;
 }
 
 /*===========================================================================*/
@@ -488,19 +546,25 @@ kvs::ColorMap ColorMap::Magma( const size_t resolution )
 /*===========================================================================*/
 kvs::ColorMap ColorMap::Cividis( const size_t resolution )
 {
-    std::list<kvs::RGBColor> colors;
-    colors.push_back( kvs::RGBColor( 255, 233,  69 ) );
-    colors.push_back( kvs::RGBColor( 228, 206,  91 ) );
-    colors.push_back( kvs::RGBColor( 197, 181, 107 ) );
-    colors.push_back( kvs::RGBColor( 166, 156, 117 ) );
-    colors.push_back( kvs::RGBColor( 137, 133, 120 ) );
-    colors.push_back( kvs::RGBColor( 111, 112, 115 ) );
-    colors.push_back( kvs::RGBColor(  85,  91, 108 ) );
-    colors.push_back( kvs::RGBColor(  54,  70, 107 ) );
-    colors.push_back( kvs::RGBColor(   0,  50, 110 ) );
-    colors.push_back( kvs::RGBColor(   0,  32,  76 ) );
+    std::list<kvs::RGBColor> colors = {
+        { 255, 233,  69 },
+        { 228, 206,  91 },
+        { 197, 181, 107 },
+        { 166, 156, 117 },
+        { 137, 133, 120 },
+        { 111, 112, 115 },
+        {  85,  91, 108 },
+        {  54,  70, 107 },
+        {   0,  50, 110 },
+        {   0,  32,  76 }
+    };
     colors.reverse();
-    return ::Interpolate( colors, resolution, kvs::ColorMap::LabSpace );
+
+    kvs::ColorMap cmap( resolution );
+    cmap.setPoints( colors );
+    cmap.setColorSpaceToLab();
+    cmap.create();
+    return cmap;
 }
 
 /*==========================================================================*/
@@ -574,6 +638,24 @@ void ColorMap::addPoint( const float value, const kvs::HSVColor color )
     this->addPoint( value, kvs::RGBColor( color ) );
 }
 
+void ColorMap::setPoints( const std::list<kvs::RGBColor>& colors )
+{
+    const size_t ncolors = colors.size();
+    const float stride = 1.0f / ( ncolors - 1 );
+
+    this->addPoint( 0.0f, colors.front() ); // start point
+    auto color = colors.begin(); color++;
+    auto end = colors.end(); end--;
+    size_t index = 1;
+    while ( color != end )
+    {
+        const float value = kvs::Math::Round( m_resolution * stride * index );
+        this->addPoint( value, *color );
+        color++; index++;
+    }
+    this->addPoint( float( m_resolution - 1 ), colors.back() ); // end point
+}
+
 /*===========================================================================*/
 /**
  *  @brief  Removes the constrol point associated with the given scalar value.
@@ -644,6 +726,12 @@ void ColorMap::create()
                         const kvs::RGBColor c0 = p0.second;
                         const kvs::RGBColor c1 = p1.second;
                         color = kvs::RGBColor::Mix( c0, c1, ratio );
+                    }
+                    else if ( m_color_space == HCLSpace )
+                    {
+                        const kvs::HCLColor c0 = p0.second;
+                        const kvs::HCLColor c1 = p1.second;
+                        color = kvs::HCLColor::Mix( c0, c1, ratio );
                     }
                     else if ( m_color_space == HSLSpace )
                     {